Basic hardware abstractions (#6)
Provide the basic hardware abstractions used:
- Binary input pins
- Binary output pins
- PWM Channels
- Hardware provider

In addition to these, define the concept of a `Registrar` which can be used to access the hardware providers (although it will _not_ own them). With these, the concept of a `HardwareManager` can be defined.

Since the binary input pin abstract contains some code, provide a mock version and tests through it. Tests for the `Registrar` are also provided.