Infrastructure for permanent way items (#10)
Create the basic infrastructure for permanent way items. The two principal classes for this are
- `PWItemModel` for managing the hardware state
- `PWItemController` for managing a thread for the `PWItemModel`

Since the `PWItemModel` is little more than an abstraction (with a few very basic methods for handling communication with the `PWItemController` there are no tests for it. Test coverage is provided by those for the `PWItemController` itself.