Skeleton for RTCClient (#17)
Working on a client for communicating track circuit changes back to rail traffic control. At the base of this is the new `RTCClient` abstraction. These clients are created by a `SoftwareManager` abstraction.

As a result of this, the `Construct` method of the data classes for permanent way items has acquired an argument to supply a `SoftwareManager`.

Provide mocks and update existing tests. Nothing actually uses the new functionality yet.