Start work on pigpiod integration (#30)
This represents the initial integration of the [`pigpiod` library](http://abyz.me.uk/rpi/pigpio/pigpiod.html). There are two main components:
- A set of stubs (in `pigpiodstubs.hpp`) which log calls if `pigpiod` is not available
- A `LibraryManager` which handles initialisation of the `pigpiod` library (or the stub library)

The `pigpiod` library will be kept alive so long as the `LibraryManager` singleton exists. We use a `std::shared_ptr` for this, relying on the [reference count being thread-safe](https://en.cppreference.com/w/cpp/memory/shared_ptr). So long as we only copy the `std::shared_ptr` to the singleton, this should suffice.