Make hardware unique (#36)
Have the `HardwareProvider` class return `std::unique_ptr`s to the hardware abstractions, since each should be owned by exactly one permanent way item.

A consequence of this is that the various mock providers are now 'dangerous' in the sense that the pointers they keep to the hardware allocated are raw pointers, which will become invalid if the corresponding hardware is deleted. For this reason, one of the tests for the `PWItemManager` has had to lose its checks that the destructors did turned off pins.