Allow longer wakeups for PWItemModels (#42)
Increase the maximum allowed wake up time which may be requested by `PWItemModel` instances. This is mainly for `TrackCircuitMonitor` so that it won't send so many extra updates to rail traffic control.

Unfortunately, this does slow down the test suite a lot, since there is a test which ensures that this limit is applied. At the moment, Boost.Test doesn't appear to support _excluding_ a label from a test run.