Install libjson-rpc-cpp for Azure DevOps (#49)
Add installation of [`libjson-rpc-cpp`](https://github.com/cinemast/libjson-rpc-cpp) to the Azure DevOps pipeline. Since the version available through `apt-get` is only v0.7.0, the script clones and builds from GitHub instead.

One of the validation tests fails, but since this is the `redis` one, this is not considered a problem.