Abstract CabinetService (#53)
Make the `CabinetService` an abstract interface, and take what was previously written as an implementation of the interface. Then, the `RPCServer` itself only has a dependency on the interface, and we can avoid dragging all the the details of the permanent way items into the`RPCServer`.

Also, define a custom `CabinetServiceResponse` type, which is a single valued `enum class`. This needed because JSON-RPC notifications can't even return errors. By returning a type (which will always indicated success), we ensure that if an exception is thrown, it will be caught and returned to the user as a server error, and not silently lost.