Update to RPC service interface (#54)
Update the abstract RPC service interface to have 'typed' and 'string' versions. This is because JSON-RPC-CXX has trouble with automatically parsing the `enum class` types. The 'string' versions take the arguments as `std::string`, convert to the required types, and then call the 'typed' version. The method names as suffixed with `String` because getting pointers to overloaded functions is tedious.

Also add an overload to parse a `TurnoutState` from a string, and correct the `curl` call for setting a multiaspect signal (was silently failing).