Add static Parse for ItemId (#55)
Add a template specialisation for `Lineside::Parse<>` for the `ItemId` type, so that strings can be parsed directly without having to create an `ItemId` object first (trying to create a static `ItemId::Parse()` method fails due to the member method with the same signature). Insert this new function into relevant places.