Implement PCA9685 device (#58)
Implement a PCA9685 device (an I2C controlled PWM channel array) and PWM channel controllers obtained from it. This builds on the `I2CDevice` in the `PiGPIOdpp` library, and adds a _different_ `I2CDevice` in the `Lineside` library. Have some basic tests as well, which are disabled if running on a real Pi.

This is just the library work; actual applications will come later.