Update JSON-RPC service (#63)
The library `json-rpc-cxx` has had an update to make it work with `enum class` types properly. Bring in update to the JSON library underlying the RPC server library. Also, add a type mapper for `ItemId` (have sent further enquiries as to whether this should be necessary). With these in place, we no longer need the `CabinetService` 'string' methods.