Create Tendril library for hardware abstraction (#72)
Create a new library, Tendril, to handle hardware abstraction. The current split between `pigpiodpp` and `lineside` is not very good. This library should provide common concepts, which are used by `lineside` and implemented by `pigpiodpp`. The only things which should 'know' about all three libraries are the actual applications.

This changeset just creates the library. Moving the other two libraries over to it will be another large changeset - which will probably reveal a variety of oversights and omissions (for example, configurations and instantiation are a little poorly defined right now). However, the needed core concepts should all be present.