Switch to Tendril library (#73)
Switch `liblineside` and `libpigpiod` to using `libtendril`. A consequence of this is that `liblineside` no longer has a dependency on `libpigpiod`.

The `DirectDriveMASH` was adjusted so that the pins are public members, to simplify testing. However, the tests of `DirectDriveMASH` are currently only 'happy path' - this will need to be corrected in future.

Furthermore, the `apps` directory has been unhooked from the build, pending further thought about how best to instantiate things in the new paradigm. However, the changeset is already large enough.