Enable HardwareManager creation (#74)
This changeset provides a cleaner interface for creating and populating `HardwareManager` instances. There doesn't appear to be a fully opaque way of doing this, so the ultimate consumer of `libtendril` has to know which implementation it wants to use (only `libpigpiodpp` available right now). This has resulted in some data transfer objects being moved over to `libtendril`. All the `apps/` have been updated to use the new paradigm.

Closes #66