Create BOPArrayMASH (#81)
Implement a multiple aspect signal head using a `BOPArray`, rather than directly driving from the individual pins. At present the only implementation is with the GPIO device, which means that there is little practical difference. However, if shift registers are being used, then this will offer a much more efficient means to operate the signal heads.

As part of this, some refactoring of the existing `DirectDriveMASH` and `MultiAspectSignalHead` has occurred, with the result that `MultiAspectSignalHead` is no longer a pure interface.

Closes #77