Basic implementation of SN74x595 Shift Register (#83)
Putting together a class to control an SN74x595 shift register (or chain thereof) which is directly controlled through GPIO pins (as opposed to using SPI). There is a small program `dd595` in the `libpigpiodpp/apps` directory which allows for testing with actual hardware.

This is not yet available from `liblineside` since I have realised that `DeviceRequestData` is going to need to live inside Tendril in order for the requests for pins to be made cleanly. That is more refactoring that I want for this changeset. Beyond that, this is very much an initial implementation. The abstractions are almost certainly not in the right places, and will have to be changed as other devices (or even just an SPI version of the SN74x595) are added.