Constructing DirectDriveSN74x595 (#85)
Starting work on the construction of a `DirectDriveSN74x595` from its corresponding 'data' object. This has prompted a rethink of how devices are handled within the libraries, and the effects are not fully in place.

Because the new device needs to be constructed from GPIO pins, it has to be passed a `HardwareManager`. For consistency, I2C devices should be constructed in the same way, with the `HardwareManager` extended with a `Registrar` for providers of `I2CCommunicator` objects. This changeset puts a lot of the required machinery in place, but doesn't switch over the `PCA9685` class to use the new paradigm. Doing that is going to ripple as far as the XML configuration files for Lineside itself, so it can go into another changeset.