New way of working with Devices (#87)
Reorganise how external devices (such as the PCA9685 and 74x595 shift register) are configured and handled. They are all now gathered together in a single `<Devices>` section in the configuration, and their 'Data' classes are responsible for their manufacture. Since devices which are purely I2C are likely to be common, have a template 'Data' class for handling these.

One particular note: I2C communicators are now configured as a `HardwareRequest` with the bus being identified through the `providerName` and the address (in hex) through the `idOnProvider` (at least within PiGPIOdpp - other implementations would be free to change this).

The changes ripple throughout the libraries and applications.